app/requirements.txt
elastic-transport==8.4.1
elasticsearch==8.10.1
elasticsearch-dsl==8.9.0
awswrangler==2.17.0
python-dotenv==1.0.0
streamlit==1.27.2
pandas==1.3.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
apache-airflow-client==2.7.2